createEncounter refactor
